# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Your API Title",
        "description": "Your API Description",
        "contact": {
            "email": "contact@example.com"
        },
        "license": {
            "name": "License Name",
            "url": "http://www.example.com/license"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/diet/plan": {
            "get": {
                "summary": "Retrieve all plans",
                "operationId": "f02cb4b6702cddb539382c49397a9d8c",
                "responses": {
                    "200": {
                        "description": "Plans retrieved successfully"
                    }
                }
            },
            "post": {
                "summary": "Create a new plan",
                "operationId": "5365fcaff3d5ab04c374599246c70ad7",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Plan's name",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Plan's status",
                        "required": false,
                        "schema": {
                            "type": "boolean"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Plan created successfully"
                    },
                    "400": {
                        "description": "Validation errors"
                    }
                }
            }
        },
        "/api/diet/plan/{plan}": {
            "get": {
                "summary": "Retrieve a plan",
                "operationId": "11e127d0f90f966cc66e8dfd8c108333",
                "responses": {
                    "200": {
                        "description": "Plan retrieved successfully"
                    },
                    "404": {
                        "description": "Plan not found"
                    }
                }
            },
            "put": {
                "summary": "Update an existing plan",
                "operationId": "88b1ab6af59b61cd8b86aae31dd76668",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Plan's name",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Plan's status",
                        "required": false,
                        "schema": {
                            "type": "boolean"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Plan updated successfully"
                    },
                    "400": {
                        "description": "Validation errors"
                    }
                }
            },
            "delete": {
                "summary": "Delete a plan",
                "operationId": "a8c44dbacfbea174551c1c321dbadb62",
                "responses": {
                    "200": {
                        "description": "Plan deleted successfully"
                    },
                    "404": {
                        "description": "Plan not found"
                    }
                }
            }
        },
        "/api/diet/plan/assign": {
            "post": {
                "summary": "Assign a plan to multiple users",
                "operationId": "9adad811ae5a4a2e4d0381ef30a3d7ef",
                "parameters": [
                    {
                        "name": "user_ids",
                        "in": "query",
                        "description": "Array of User IDs",
                        "required": true,
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "integer"
                            }
                        }
                    },
                    {
                        "name": "plan_id",
                        "in": "query",
                        "description": "Plan's ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Plan assigned to users successfully"
                    },
                    "400": {
                        "description": "Validation errors"
                    }
                }
            }
        },
        "/api/plans": {
            "post": {
                "tags": [
                    "Plans"
                ],
                "summary": "Create a new full plan",
                "description": "Create a new full plan with meals and items",
                "operationId": "createFullPlan",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Plan Name"
                                    },
                                    "note": {
                                        "type": "string",
                                        "example": "This is a plan note"
                                    },
                                    "meals": {
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "name": {
                                                    "type": "string",
                                                    "example": "Meal 1"
                                                },
                                                "note": {
                                                    "type": "string",
                                                    "example": "This is a meal note"
                                                },
                                                "items": {
                                                    "type": "array",
                                                    "items": {
                                                        "properties": {
                                                            "name": {
                                                                "type": "string",
                                                                "example": "Item 1"
                                                            },
                                                            "type": {
                                                                "type": "integer",
                                                                "example": 0
                                                            },
                                                            "details": {
                                                                "type": "array",
                                                                "items": {
                                                                    "properties": {
                                                                        "name": {
                                                                            "type": "string",
                                                                            "example": "Detail Name"
                                                                        },
                                                                        "number": {
                                                                            "type": "integer",
                                                                            "example": 1
                                                                        },
                                                                        "standard_type": {
                                                                            "type": "integer",
                                                                            "example": 1
                                                                        },
                                                                        "carbohydrate": {
                                                                            "type": "number",
                                                                            "format": "float",
                                                                            "example": 30
                                                                        },
                                                                        "protein": {
                                                                            "type": "number",
                                                                            "format": "float",
                                                                            "example": 20
                                                                        },
                                                                        "fat": {
                                                                            "type": "number",
                                                                            "format": "float",
                                                                            "example": 10
                                                                        },
                                                                        "calories": {
                                                                            "type": "number",
                                                                            "format": "float",
                                                                            "example": 100
                                                                        }
                                                                    },
                                                                    "type": "object"
                                                                }
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "user_ids": {
                                        "type": "array",
                                        "items": {
                                            "type": "integer"
                                        },
                                        "example": [
                                            1,
                                            2,
                                            3
                                        ]
                                    },
                                    "is_work": {
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Plan created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Plan created successfully"
                                        },
                                        "plan": {
                                            "$ref": "#/components/schemas/PlanResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "ItemDetailResource": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 50
                    },
                    "item_id": {
                        "type": "integer",
                        "example": 40
                    },
                    "item": {
                        "type": "string",
                        "example": "Item Name"
                    },
                    "name": {
                        "type": "string",
                        "example": "Detail Name"
                    },
                    "calories": {
                        "type": "number",
                        "format": "float",
                        "example": 0
                    },
                    "standard_name": {
                        "type": "string",
                        "example": "Standard Name"
                    },
                    "carbohydrate": {
                        "type": "number",
                        "format": "float",
                        "example": 30
                    },
                    "protein": {
                        "type": "number",
                        "format": "float",
                        "example": 20
                    },
                    "fat": {
                        "type": "number",
                        "format": "float",
                        "example": 10
                    }
                },
                "type": "object"
            },
            "ItemResource": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 40
                    },
                    "name": {
                        "type": "string",
                        "example": "Item 1"
                    },
                    "calories": {
                        "type": "number",
                        "format": "float",
                        "example": 0
                    },
                    "type": {
                        "type": "string",
                        "example": "recipe"
                    },
                    "standard_name": {
                        "type": "string",
                        "example": "1"
                    },
                    "number": {
                        "type": "integer",
                        "example": 1
                    },
                    "carbohydrate": {
                        "type": "number",
                        "format": "float",
                        "example": 30
                    },
                    "fat": {
                        "type": "number",
                        "format": "float",
                        "example": 10
                    },
                    "protein": {
                        "type": "number",
                        "format": "float",
                        "example": 20
                    },
                    "item_details": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/ItemDetailResource"
                        }
                    },
                    "has_details": {
                        "type": "boolean",
                        "example": true
                    }
                },
                "type": "object"
            },
            "MealResource": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 37
                    },
                    "name": {
                        "type": "string",
                        "example": "Meal 1"
                    },
                    "calories": {
                        "type": "number",
                        "format": "float",
                        "example": 250
                    },
                    "fat": {
                        "type": "number",
                        "format": "float",
                        "example": 10
                    },
                    "carbohydrate": {
                        "type": "number",
                        "format": "float",
                        "example": 30
                    },
                    "protein": {
                        "type": "number",
                        "format": "float",
                        "example": 20
                    },
                    "items_count": {
                        "type": "integer",
                        "example": 2
                    },
                    "note": {
                        "type": "string",
                        "example": "this note is optional called meal note"
                    },
                    "items": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/ItemResource"
                        }
                    },
                    "is_eaten": {
                        "type": "boolean",
                        "example": false
                    }
                },
                "type": "object"
            },
            "PlanResource": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 27
                    },
                    "name": {
                        "type": "string",
                        "example": "Plan Name"
                    },
                    "total_calories": {
                        "type": "number",
                        "format": "float",
                        "example": 500
                    },
                    "total_carbohydrate": {
                        "type": "number",
                        "format": "float",
                        "example": 60
                    },
                    "total_protein": {
                        "type": "number",
                        "format": "float",
                        "example": 40
                    },
                    "total_fat": {
                        "type": "number",
                        "format": "float",
                        "example": 20
                    },
                    "notes": {
                        "type": "string",
                        "example": "this note is optional called plan note"
                    },
                    "meals": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/MealResource"
                        }
                    }
                },
                "type": "object"
            },
...
        }
    }
}